+++ build.gradle
@@ -54,7 +54,6 @@ dependencies {
 
     // libs included in the final jar
     api group: 'org.apache.pulsar', name: 'pulsar-client', version: '3.1.2'
-    api group: 'org.apache.pulsar', name: 'pulsar-client-admin', version: '3.1.2'
 }
 
 
@@ -87,6 +86,8 @@ dependencies {
     testImplementation platform("io.micronaut.platform:micronaut-platform:$micronautVersion")
     testImplementation "io.micronaut.test:micronaut-test-junit5"
     testImplementation "io.micronaut.reactor:micronaut-reactor"
+
+    testImplementation group: 'org.apache.pulsar', name: 'pulsar-client-admin', version: '3.1.2'
 
     // test deps needed only for to have a runner
     testImplementation group: "io.kestra", name: "core", version: kestraVersion
