+++ gradle.properties
@@ -2,4 +2,3 @@ version=0.15.0-SNAPSHOT
 kestraVersion=[0.15,)
 micronautVersion=4.3.0
 lombokVersion=1.18.30
-confluentVersion=7.2.1
