+++ src/main/java/io/kestra/plugin/pulsar/SchemaType.java
@@ -1,12 +1,7 @@
 package io.kestra.plugin.pulsar;
 
 public enum SchemaType {
-    // No Schema
     NONE,
-    
-    // Avro schems
     AVRO,
-    
-    // JSON Schema
     JSON
 }
