+++ src/main/java/io/kestra/plugin/pulsar/Trigger.java
@@ -80,16 +80,16 @@ public class Trigger extends AbstractTrigger implements PollingTriggerInterface,
 
     private String consumerName;
 
-    @io.swagger.v3.oas.annotations.media.Schema(
-      title = "JSON string of the topic's schema",
-      description = "Required for connecting with topics with a defined schema and strict schema checking"
+    @Schema(
+        title = "JSON string of the topic's schema",
+        description = "Required for connecting with topics with a defined schema and strict schema checking"
     )
     @PluginProperty(dynamic = true)
     protected String schemaString;
 
-    @io.swagger.v3.oas.annotations.media.Schema(
-      title = "The schema type of the topic",
-      description = "Can be one of either AVRO or JSON. Leave as NONE for topics without strict schema checking"
+    @Schema(
+        title = "The schema type of the topic",
+        description = "Can be one of NONE, AVRO or JSON. None means there will be no schema enforced."
     )
     @PluginProperty(dynamic = true)
     @Builder.Default
